move files to the place where prove expects them
I like to run my tests with prove. This commit moves PDL.pm and is_pdl.t
to the place where prove expects them to be. A .proverc file is included
with this commit. The major advantage of this commit is that prove will
now run the tests against the development version of Test::PDL, instead
of a version of Test::PDL that happens to be installed on your system.
(Not a fan of prove? perl -Ilib t/is_pdl.t also tests against the
development version.)