fix =for commands in documentation
After the fix by David Mertens, which corrected a bug introduced in an
earlier commit and which made the PDL docs database scraper work again
on Test::PDL, Dist::Zilla started writing empty =begin/=end regions for
the empty =for commands. This commands fixes that problem by adding a
comment after the =for commands, making both PDL and Dist::Zilla happy.
The PDL docs database scraper doesn't seem to mind, and continues to
find the documentation.