Use per-invocation options instead of global variables
Up to now, the comparison code used configurable parameters that could
only be changed globally. Now they can be overridden per invocation,
e.g., as in

	is_pdl($got, $expected, { atol => 1e-3 });

This leads to fewer surprises and arguably better code. It is still
possible to specify the default comparison parameters at import time.

There were too many ways to set the comparison parameters: at import
time, via set_options(), and per invocation. Moreover, the pervasive
effect of set_options() was a potential source of errors. Note that it
is still possible, though not encouraged or supported (or even
documented) to localize changes to %DEFAULTS with something like:

	{
		local $Test::PDL::DEFAULTS{atol} = 1e-3;
		...
	}