Use per-invocation options instead of global variables
Up to now, the comparison code used configurable parameters that could
only be changed globally. Now they can be overridden per invocation,
e.g., as in

	is_pdl($got, $expected, { TOLERANCE => 1e-3 });

This leads to less surprises and arguably better code. It is still
possible to specify the default comparison parameters at import time,
and with the function 'set_defaults', but the latter is about to
disappear.