Remove set_defaults altogether
There were too many ways to set the comparison parameters: at import
time, via set_defaults(), and per invocation. Moreover, the pervasive
effect of set_defaults() was a potential source of errors. Note that it
is still possible, though not encouraged or supported (or even
documented) to localize changes to %DEFAULTS with something like:

	{
		local $Test::PDL::DEFAULTS{TOLERANCE} = 1e-3;
		...
	}